backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "interviewcraft-backend"
version = "0.1.0"
description = "Deliberate Practice Engine for Interviews — Backend"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.0",
    "structlog>=24.0.0",
    "anthropic>=0.40.0,<1.0.0",
    "openai>=1.0.0,<2.0.0",
    "pipecat-ai[deepgram,anthropic,elevenlabs]>=0.0.76,<1.0.0",
    "numpy>=2.0.0,<3.0.0",   # pipecat-ai allows 1.26.4+ but 1.x has no py313 wheel
    "deepgram-sdk>=4.0.0,<5.0.0",
    "httpx>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=3.2.0,<4.0.0",
    "cryptography>=42.0.0",
    "authlib>=1.3.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "PyPDF2>=3.0.0",
    "python-docx>=1.0.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "types-passlib>=1.7.7",
    "types-python-jose>=3.3.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests requiring live API keys (deselect with '-m not integration')",
    "nightly: marks tests as nightly full-suite tests (30 cases × 5 runs)",
]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = false
ignore_missing_imports = true
